Separate collection query to optimize app shell
